requirements.txt
memvid>=0.1.3
PyPDF2>=3.0.0
python-dotenv>=0.19.0
opencv-python>=4.5.0
